lib: move SAML timestring utils into lib
